wip increasing token count by 100

+++ plugins/tff_backend/models/global_stats.py
@@ -33,7 +33,7 @@ class GlobalStats(NdbModel):
     name = ndb.StringProperty()
     token_count = ndb.IntegerProperty()
     unlocked_count = ndb.IntegerProperty()
-    value = ndb.IntegerProperty()  # Value in dollar
+    value = ndb.FloatProperty()  # Value in dollar
     # Value per other currency
     currencies = ndb.LocalStructuredProperty(CurrencyValue, repeated=True)  # type: list[CurrencyValue]
     market_cap = ndb.ComputedProperty(lambda self: self.value * self.unlocked_count, indexed=False)
